chore: restructure sidebar components
